support TLS options
