+++ src/aws/aws_sdk/mod.rs
@@ -1,7 +1,7 @@
 use crate::aws::AwsAuthentication;
 use crate::config::ProxyConfig;
-use crate::http::build_proxy_connector;
-use crate::tls::MaybeTlsSettings;
+use crate::http::{build_proxy_connector, build_tls_connector};
+use crate::tls::{MaybeTlsSettings, TlsOptions};
 use aws_smithy_client::erase::DynConnector;
 use aws_smithy_client::SdkError;
 use aws_smithy_http::endpoint::Endpoint;
@@ -72,8 +72,6 @@ pub trait ClientBuilder {
 
     fn client_from_conf_conn(builder: Self::ConfigBuilder, connector: DynConnector)
         -> Self::Client;
-
-    fn client_from_conf(builder: Self::ConfigBuilder) -> Self::Client;
 }
 
 pub async fn create_client<T: ClientBuilder>(
@@ -81,6 +79,7 @@ pub async fn create_client<T: ClientBuilder>(
     region: Option<Region>,
     endpoint: Option<Endpoint>,
     proxy: &ProxyConfig,
+    tls_options: &Option<TlsOptions>,
 ) -> crate::Result<T::Client> {
     let mut config_builder = T::create_config_builder(auth.credentials_provider().await?);
 
@@ -92,13 +91,17 @@ pub async fn create_client<T: ClientBuilder>(
         config_builder = T::with_region(config_builder, region);
     }
 
-    if proxy.enabled {
-        let tls_settings = MaybeTlsSettings::enable_client()?;
+    let tls_settings = MaybeTlsSettings::tls_client(tls_options)?;
+
+    let connector = if proxy.enabled {
         let proxy = build_proxy_connector(tls_settings, proxy)?;
         let hyper_client = aws_smithy_client::hyper_ext::Adapter::builder().build(proxy);
-        let connector = aws_smithy_client::erase::DynConnector::new(hyper_client);
-        Ok(T::client_from_conf_conn(config_builder, connector))
+        aws_smithy_client::erase::DynConnector::new(hyper_client)
     } else {
-        Ok(T::client_from_conf(config_builder))
-    }
+        let tls_connector = build_tls_connector(tls_settings)?;
+        let hyper_client = aws_smithy_client::hyper_ext::Adapter::builder().build(tls_connector);
+        aws_smithy_client::erase::DynConnector::new(hyper_client)
+    };
+
+    Ok(T::client_from_conf_conn(config_builder, connector))
 }
