+++ src/common/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(any(feature = "sources-datadog_agent", feature = "sinks-datadog_metrics"))]
 pub(crate) mod datadog;
+
+#[cfg(any(feature = "sources-aws_sqs", feature = "sinks-aws_sqs"))]
+pub(crate) mod sqs;
